+++ Bizap/app/src/main/java/com/emul8r/bizap/data/local/entities/PrefilledItemEntity.kt
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey
 data class PrefilledItemEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val description: String,
-    val unitPrice: Double
+    val unitPrice: Long                 // Cents (e.g., 4999 = $49.99)
 )
